Adding logging extras. Separating files based on functionality.
