+++ reusables/__init__.py
@@ -1,3 +1,6 @@
 from __future__ import absolute_import
 from .reusables import *
+from .namespace import *
+from .dt import *
+from .log import *
 from . import dangerzone
