+++ reusables/reusables.py
@@ -9,12 +9,12 @@ import os
 import sys
 import re
 import tempfile as _tempfile
-import logging as _logging
-import datetime as _datetime
-import time as _time
+
+from .namespace import Namespace
+from .log import get_logger
 
 __author__ = "Chris Griffith"
-__version__ = "0.2.0"
+__version__ = "0.3.0"
 
 python_version = sys.version_info[0:3]
 version_string = ".".join([str(x) for x in python_version])
@@ -25,10 +25,7 @@ nix_based = os.name == "posix"
 win_based = os.name == "nt"
 temp_directory = _tempfile.gettempdir()
 
-logger = _logging.getLogger(__name__)
-if python_version >= (2, 7):
-    # Suppresses warning that no logger is found if a parent logger is not set
-    logger.addHandler(_logging.NullHandler())
+logger = get_logger(__name__)
 
 # http://msdn.microsoft.com/en-us/library/aa365247%28v=vs.85%29.aspx
 
@@ -68,39 +65,6 @@ reg_exps = {
                              r'[\. \-]?\d{4})')
         }
     },
-    "datetime": {
-        "format": {
-            "%I": re.compile(r"\{(?:12)?\-?hours?\}"),
-            "%H": re.compile(r"\{24\-?hours?\}"),
-            "%S": re.compile(r"\{seco?n?d?s?\}"),
-            "%M": re.compile(r"\{minu?t?e?s?\}"),
-            "%f": re.compile(r"\{micro\-?(?:second)?s?\}"),
-            "%Z": re.compile(r"\{(?:(tz|time\-?zone))?\}"),
-            "%y": re.compile(r"\{years?\}"),
-            "%Y": re.compile(r"\{years?\-?(?:(full|name|full\-?name))?s?\}"),
-            "%m": re.compile(r"\{months?\}"),
-            "%b": re.compile(r"\{months?\-?name\}"),
-            "%B": re.compile(r"\{months?\-?(?:(full|full\-?name))?s?\}"),
-            "%d": re.compile(r"\{days?\}"),
-            "%w": re.compile(r"\{week\-?days?\}"),
-            "%j": re.compile(r"\{year\-?days?\}"),
-            "%a": re.compile(r"\{(?:week)?\-?days?\-?name\}"),
-            "%A": re.compile(r"\{(?:week)?\-?days?\-?fullname\}"),
-            "%U": re.compile(r"\{weeks?\}"),
-            "%W": re.compile(r"\{mon(?:day)?\-?weeks?\}"),
-            "%x": re.compile(r"\{date\}"),
-            "%X": re.compile(r"\{time\}"),
-            "%c": re.compile(r"\{date\-?time\}"),
-            "%z": re.compile(r"\{(?:utc)?\-?offset\}"),
-            "%p": re.compile(r"\{periods?\}"),
-            "%Y-%m-%dT%H:%M:%S": re.compile(r"\{iso-?(?:format)?\}")
-        },
-        "date": re.compile(r"((?:[\d]{2}|[\d]{4})[\- _\\/]?[\d]{2}[\- _\\/]?"
-                           r"\n[\d]{2})"),
-        "time": re.compile(r"([\d]{2}:[\d]{2}(?:\.[\d]{6})?)"),
-        "datetime": re.compile(r"((?:[\d]{2}|[\d]{4})[\- _\\/]?[\d]{2}[\- _\\/]"
-                               r"?[\d]{2}T[\d]{2}:[\d]{2}(?:\.[\d]{6})?)")
-    }
 }
 
 common_exts = {
@@ -130,99 +94,6 @@ common_variables = {
         },
     },
 }
-
-
-class Namespace(dict):
-    """
-    Namespace container.
-    Allows access to attributes by either class dot notation or item reference
-
-    All valid:
-        namespace.spam.eggs
-        namespace['spam']['eggs']
-        namespace['spam'].eggs
-    """
-
-    def __init__(self, *args, **kwargs):
-        if len(args) == 1 and isinstance(args[0], dict):
-            kwargs = args[0]
-        for k, v in kwargs.items():
-            if isinstance(v, dict):
-                v = Namespace(v)
-            setattr(self, k, v)
-
-    def __contains__(self, item):
-        try:
-            return dict.__contains__(self, item) or hasattr(self, item)
-        except Exception:
-            return False
-
-    def __getattr__(self, item):
-        try:
-            return object.__getattribute__(self, item)
-        except AttributeError:
-            try:
-                return self[item]
-            except KeyError:
-                raise AttributeError(item)
-
-    def __setattr__(self, key, value):
-        if isinstance(value, dict):
-            value = Namespace(**value)
-        try:
-            object.__getattribute__(self, key)
-        except AttributeError:
-            try:
-                self[key] = value
-            except Exception:
-                raise AttributeError(key)
-        else:
-            object.__setattr__(self, key, value)
-
-    def __delattr__(self, item):
-        try:
-            object.__getattribute__(self, item)
-        except AttributeError:
-            try:
-                del self[item]
-            except KeyError:
-                raise AttributeError(item)
-        else:
-            object.__delattr__(self, item)
-
-    def __repr__(self):
-        return "<Namespace: {0}...>".format(str(self.to_dict())[0:32])
-
-    def __str__(self):
-        return str(self.to_dict())
-
-    @classmethod
-    def from_dict(cls, dictionary):
-        if not isinstance(dictionary, dict):
-            raise TypeError("Must be a dictionary")
-        return cls(dictionary)
-
-    def to_dict(self, in_dict=None):
-        in_dict = in_dict if in_dict else self
-        out_dict = dict()
-        for k, v in in_dict.items():
-            if isinstance(v, Namespace):
-                v = v.to_dict()
-            out_dict[k] = v
-        return out_dict
-
-    def tree_view(self, sep="    "):
-        base = self.to_dict()
-        return tree_view(base, sep=sep)
-
-
-def tree_view(dictionary, level=0, sep="|  "):
-    """
-    View a dictionary as a tree.
-    """
-    return "".join(["{0}{1}\n{2}".format(sep * level, k,
-                   tree_view(v, level + 1, sep=sep) if isinstance(v, dict)
-                   else "") for k, v in dictionary.items()])
 
 
 def os_tree(directory):
@@ -640,62 +511,6 @@ def extract_all(archive_file, path=".", dnd=True, enable_rar=False):
     if not dnd:
         logger.debug("Archive {0} will now be deleted".format(archive_file))
         os.unlink(archive_file)
-
-
-class DateTime(_datetime.datetime):
-
-    def __new__(cls, year=None, month=None, day=None, hour=0, minute=0,
-                second=0, microsecond=0, tzinfo=None):
-        #  Taken from datetime.datetime.now()
-        if year is not None:
-            return super(DateTime, cls).__new__(cls, year, month, day, hour,
-                                                minute, second, microsecond,
-                                                tzinfo)
-        if tzinfo is not None and not isinstance(tzinfo,
-                                                 _datetime.datetime.tzinfo):
-            raise TypeError("tzinfo argument must be None or a tzinfo subclass")
-        converter = _time.localtime if tzinfo is None else _time.gmtime
-        t = _time.time()
-        t, frac = divmod(t, 1.0)
-        us = int(frac * 1e6)
-        tz = None
-        if us == 1000000:
-            t += 1
-            us = 0
-        y, m, d, hh, mm, ss, weekday, jday, dst = converter(t)
-        ss = min(ss, 59)
-        return super(DateTime, cls).__new__(cls, y, m, d, hh, mm, ss, us, tz)
-
-    def __init__(self, *args):
-        self.__dict__ = dict(
-            year=self.year, month=self.month, day=self.day, hour=self.hour,
-            minute=self.minute, second=self.second,
-            microsecond=self.microsecond, timezone=self.tzinfo)
-
-    def format(self, desired_format, *args, **kwargs):
-        for strf, exp in regex.datetime.format.items():
-            desired_format = exp.sub(strf, desired_format)
-        return self.strftime(desired_format.format(*args, **kwargs))
-
-    def __iter__(self):
-        for k, v in self.__dict__.items():
-            yield (k, v)
-
-    @classmethod
-    def from_iso(cls, iso_datetime):
-        """
-        :type iso_datetime: str
-        """
-        try:
-            assert regex.datetime.datetime.match(iso_datetime).groups()[0]
-        except (ValueError, AssertionError, IndexError, AttributeError):
-            raise TypeError("String is not in ISO format")
-        try:
-            return cls.strptime(iso_datetime, "%Y-%m-%dT%H:%M:%S.%f")
-        except ValueError:
-            return cls.strptime(iso_datetime, "%Y-%m-%dT%H:%M:%S")
-
-    # TODO add a 'from datetime'
 
 
 def main(command_line_options=""):
